DeveloperGuide: update section on UndoRedoStack to VersionedAddressBook
The app previously uses UndoRedoStack which stores the UndoableCommands
in LogicManager for the undo/redo mechanism.

Since the undo/redo machanism has been shifted to VersionedAddressBook
which stores ReadOnlyAddressBook in a List, and has been moved to Model
layer instead, the sections on undo/redo feature, Logic and Model
component in DeveloperGuide is out of date.

Also, commands that do not mutate the address book, such as
FindCommand and ListCommand, will no longer clear the redoable
address book states. This allows for redo commands to pass when we
undo commands and execute a non-mutating command afterwards.

UndoableCommand has also been removed from the codebase as well.

Let's update the out of date sections in the DeveloperGuide, such
as switching from UndoRedoStack to the updated VersionedAddressBook,
how the undo/redo mechanism works, and remove any reference
to UndoableCommand in DeveloperGuide.