HelpWindow: don't show site navigation bar
The help window launched by the `help` command displays the rendered
User Guide (UserGuide.html). However, it contains a navigation bar.

This navigation bar is non-functional -- clicking on the links will not
do anything because the files that they link to are not found. This
is because we only copy in `UserGuide.html` to the resources directory,
and not the rest of the documentation.

As such, let's hide the navigation bar when the User Guide is displayed
in the Help Window.

To do this, introduce a separate document, HelpWindow.adoc, that is only
meant to be displayed inside the help window. This document just
includes the contents of UserGuide.adoc verbatim, with one exception: it
sets the `no-site-header` attribute, thus disabling the rendering of the
navigation bar.