parsers: remove unnecessary catch-throws of ParseExceptions
These try-catch blocks simply catch the ParseException, and then throw a
new ParseException with the exact same message and the caught
ParseException as its cause.

This catch-throwing of ParseExceptions does not add any value, so let's
remove them. This has the side benefit of removing one level of
indentation from the code.