ParserUtil: remove `Optional` variants of parse methods
As the comment[1] at the top of ParserUtil says, methods that take
`Optional` as a parameter are bad.

So, let's remove them.

(It could be argued that leaving these methods in the code base could
serve as an example for students of what not to do, but really, (1) the
number of things that are forbidden under "best practice" is quite
numerous, and trying to add an example of every single one of them in
our code base would just lead to an annoyingly smelly code base, and (2)
students will need to spend time cleaning up all of these negative
examples, which is an unnecessary burden. These things would be better
documented in style guides.)

[1] Comment added in 84d6516f ([#557] ParserUtil: Update header comments
    on Optional usage (#558), 2017-07-18)