FindCommandTest: verify entire model state
FindCommandTest only verifies some parts of the model state, namely the
contents of `model.getAddressBook()` and
`model.getFilteredPersonList()`.

However, the Model contains much more than that -- it also encapsulates
the address book history, for example. The Model may also grow to store
other pieces of data in the future. We need to ensure that FindCommand
does not needlessly meddle with the other data stored in the model.

As such, let's make FindCommandTest verify the entire model state.
This is done by rewriting it to use
`CommandTestUtil.assertCommandSuccess(...)`, to make it consistent with
the other command tests.