HelpCommandTest: verify model state as well
When executed, HelpCommand should:

 (1) Raise a ShowHelpRequestEvent, and

 (2) Not modify the model at all.

HelpCommandTest verifies (1), but does not verify (2).

Given that in the actual application, LogicManager will pass the Model
to HelpCommand, we should verify that HelpCommand behaves as expected
and does not modify the model.

As such, let's make HelpCommandTest verify the model state as well. This
is done by rewriting it to use
`CommandTestUtil.assertCommandSuccess(...)`, for consistency with the
other command tests.