AddCommandSystemTest: remove outdated comment
The bug mentioned in the comment refers to this comment that was in
AddressBook.java:

    // TODO: the tags master list will be updated even though the below line fails.
    // This can cause the tags master list to have additional tags that are not tagged to any person
    // in the person list.

which was added in 1eb8b138 (teach classes in Model to execute edit,
2016-12-13).

However, the bug was fixed in d51894a8 (AddressBook: remove master tag
list, 2018-03-29).

So, this comment is not relevant any more. Let's remove it.

No other changes will need to be done to the test case itself: any tag
will be suitable for this test case, since HOON himself does not have
any tags.