model: make UserPrefs part of model state
Despite the architecture of AB-4 saying that UserPrefs is part of the
model, and UserPrefs being in the model package, UserPrefs is not
treated as part of the model state. This makes it harder to write tests
against the model, since we need to check both the Model _and_
UserPrefs.

Make UserPrefs part of the model state, so that we can reason about and
write tests against the model as one single entity.