ModelManager: keep its own copy of UserPrefs
ModelManager keeps a reference to the UserPrefs object that is passed
to it via its constructor or via ModelManager#setUserPrefs(UserPrefs).

However, other objects aside from ModelManager may also hold a reference
to that UserPrefs object, and modify it. If developers are not careful,
this could be a source of bugs.

As such, let's teach ModelManager to keep its own defensive copy of the
UserPrefs object.