Keep a strong reference to created unmodifiableObservableLists
The observable list created with
FXCollections.unmodifiableObservableList(...) installs a "weak listener"
-- that is, the backing list only keeps a weak reference to the
unmodifiable list. This means that if there is no other strong reference
pointing to the created unmodifiable observable list, it would be
garbage collected.

This becomes a problem when other code attaches event listeners to the
unmodifiable list, with the expectation that it would be called whenever
the backing list changes. When the unmodifiable list is garbage
collected, the event listeners will not be called any more, which could
cause incorrect application behavior.

Fix this by making the code base keep a strong reference to each
unmodifiable observable list, and return the same unmodifiable
observable list whenever its getter is called.