ui: move command execution responsibility from CommandBox to MainWindow
CommandBox is the UiPart responsible for command execution (i.e calling
Logic#execute(String) and processing its return value).

In the next few commits, we want the methods of other UiParts to be
called when command execution occurs.

However, CommandBox does not have access to the other UiParts in the
main window. (And it should not, to be consistent with our architecture
design).

On the other hand, MainWindow _does_ have access to the other UiParts in
the MainWindow.

As such, let's make MainWindow responsible for command execution.
CommandBox will use callbacks to notify MainWindow that it should
execute a command.

Since now MainWindow, instead of CommandBox, is the class that directly
calls Logic#execute(String), update the Ui class diagram to reflect this
updated relationship.