BrowserPanel#DEFAULT_PAGE: provide full URL
BrowserPanel#DEFAULT_PAGE has the value of "default.html". However, the
value itself doesn't give us the full picture since it does not tell us
_where_ exactly "default.html" is located.

As such, our test code has to duplicate this information in various
places:

    MainApp.class.getResource(FXML_FILE_FOLDER + DEFAULT_PAGE)

(namely, that "default.html" is inside the FXML_FILE_FOLDER, and
FXML_FILE_FOLDER is relative to MainApp's package)

Remove this duplication of information by making DEFAULT_PAGE provide
the _full_ URL, so that we can DRY up the test code.

At the same time, wrap the initialization of DEFAULT_PAGE with a
requireNonNull(...), so that we can detect as early as possible whether
the "default.html" page is missing, which should help in debugging.