StringUtilTest: rename isUnsignedPositiveInteger() to isNonZeroUnsignedInteger()
Currently, StringUtilTest has a method named
isUnsignedPositiveInteger. However, the method tested is actually
named isNonZeroUnsignedInteger. This can cause confusion in the
naming.

It appears that, as early as #85 [1], the tested method was originally
isUnsignedInteger(). However, the test method was named
isUnsignedPositiveInteger() and was incorrect.
Additionally, in #460 [2], the tested method was renamed to
isNonZeroUnsignedInteger(), however the test was not renamed then.

To maintain consistent naming convention, lets rename the test method
to isNonZeroUnsignedInteger().

[1]: https://github.com/se-edu/addressbook-level4/commit/92bc32b631c62041625a021719adf7c280711570
[2]: https://github.com/se-edu/addressbook-level4/commit/0352e6b6e3f51910edbff7a786f6ae24f453d38f