Import the original AB-3 LearningOutcomes
As we are planning to replace AB-3 with this repo, let's import the
original AB-3 Learning Outcomes document and diagrams from the AB-3 repo
(as of commit 841298aa618dbffe710469639ed922c27752cd36).

For now, we will import the document as-is, even though the source code
in this repo has some differences compared to the original AB-3, and so
some of the Learning Outcomes exercises can't be directly applied. We
will update the Learning Outcomes document incrementally in later PRs.

The original authors of the Learning Outcomes document are listed below
(in alphabetical order):

Co-authored-by: Damith C. Rajapakse <damith@gmail.com>
Co-authored-by: Jun An <E0002937@u.nus.edu>
Co-authored-by: Paul Tan <pyokagan@gmail.com>
Co-authored-by: Tan Wang Leng <tanwangleng@outlook.com>