LO-Interfaces: update to match the current codebase
LO-Interfaces references ReadOnlyPerson, which was removed in 345d05f6
(Remove ReadOnlyPerson, 2018-01-10).

Update it to the current codebase by replacing it with
AddressBook/ReadOnlyAddressBook. Conveniently, it also does the whole
"clients who don't need write access to Foo can use ReadOnlyFoo" thing,
so the LO text doesn't need to change much.

However, unlike ReadOnlyPerson, the UI in the current codebase doesn't
depend on ReadOnlyAddressBook, since it is not concerned with the
address book at a whole, but rather its individual components (e.g. the
person list) which are forwarded individually by the logic layer. Only
the Storage and Logic layers are interested in ReadOnlyAddressBook. As
such, let's replace "MainWindow" in the diagram with "StorageManager" so
it matches the current code.

(While we are here, also update the gradients/colors used in the diagram
to match the ones used in the developer guide class diagrams.)