docs: tell readers that the screenshots are meant for AB-4
As the AB-3.5 repo inherited all of AB-4's documentation, it also
inherited their screenshots as well. However, these screenshots refer to
`addressbook-level4`, for use with the AB-4 repo, instead of
`addressbook-level35` which should be used with the AB-3.5 repo.

As such, let's add a note at the top of these documents informing
readers of this fact.

Yes, this means that readers would have to mentally replace all
references to `addressbook-level4` in the screenshots with
`addressbook-level35`, which could potentially confuse them. Ideally,
these screenshots should be modified to refer to `addressbook-level35`,
and the note should be added to the AB-4 repo instead. This is because
AB3.5 comes chronologically before AB-4, and readers would likely
encounter the documents in the AB-3.5 repo first and benefit most from
having clear unambiguous screenshots.

However, now is not a good time to update all the screenshots as the
AB-3.5 repo is still in flux. As such, let's settle for the note first,
with a plan to update the screenshots in the future.