Upgrade Codecov GitHub Action to v2 (#128)
Codecov GitHub Action v1 uses a deprecated bash uploader, which may
cease to function soon. The v2 version also has a files arg instead of a
file arg, suggesting possible deprecation of the file arg.

Let's upgrade to v2, use the files arg for coverage.xml and the directory
arg for the directory in which coverage.xml is found.