Remove support for repeated parameters in commands
If a single-valued parameter is repeated multiple times in a command,
only the last occurrence is taken. This allows the user to correct a
mistake in a parameter value by simply typing it again, which may be
faster than moving the cursor back to the original parameter values and
editing it. This feature serves as a good example of 'going the extra
mile to optimize the application for target users' (i.e., users who can
type fast).
However, this feature can cause certain user errors to go unnoticed
e.g., unintended repetition caused by a typo in parameter name

Alternatives considered:
(a) Retain current behavior but give a friendly warning to the user
    that a parameter was 'overridden'.
(b) Remove this feature i.e., treat repeated single-valued parameters
    as an error, and reject the command with an error message.

Alternative (a) requires significantly more complicated code (see #181),
steepening the learning curve for for new programmers. Given that this
application is primarily meant for training novice programmers, the
impact on the learning curve is not worth the benefit of retaining this
feature.

Therefore, let's do option (b).