Fix double-space between index and name
The spacing attribute for HBox in PersonListCard.fxml is set to '5',
causing it to look like a double-space is present in the Person
List Card (e.g. 1.  John)

Let's reduce the spacing attribute of the HBox in
PersonListCard.fxml (which determines the spacing between child
elements) to 0.5, which seems to produce the desired gap, so that
it looks like a single space is present instead of double
(e.g. 1. John).