Update setup-java action to v4
The current version (v1) of the setup-java action is outdated,
and may stop working in the future. As newer versions of the action
require the 'distribution' field, 'zulu' is specified, which is
the default used in v1.

Let's update the workflow dependency, and use the format as recommended
in https://github.com/actions/setup-java/blob/main/README.md.